chore: migrate lance-migrations test to bun test with real sqlite and lancedb; add deps injection for getConnection
